apps/observe-dashboard/src/requirements.txt
streamlit>=1.50.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.24.1
databricks-sql-connector>=2.9.3
python-dotenv>=1.0.0
